[refactor] Move type related utils away from lang_util.h
