+++ taichi/ir/frontend.cpp
@@ -2,7 +2,6 @@
 
 #include "taichi/ir/frontend.h"
 #include "taichi/ir/statements.h"
-#include "taichi/program/program.h"
 
 TLANG_NAMESPACE_BEGIN
