+++ taichi/ir/frontend.h
@@ -3,7 +3,6 @@
 
 #include "taichi/common/core.h"
 #include "taichi/common/dict.h"
-#include "taichi/util/io.h"
 #include "taichi/ir/frontend_ir.h"
 
 namespace taichi {
