+++ taichi/lang_util.cpp
@@ -65,88 +65,6 @@ real measure_cpe(std::function<void()> target,
   return elasped_cycles / float64(total_batches * elements_per_call);
 }
 
-std::string data_type_format(DataType dt) {
-  if (dt->is_primitive(PrimitiveTypeID::i32)) {
-    return "%d";
-  } else if (dt->is_primitive(PrimitiveTypeID::u32)) {
-    return "%u";
-  } else if (dt->is_primitive(PrimitiveTypeID::i64)) {
-#if defined(TI_PLATFORM_UNIX)
-    return "%lld";
-#else
-    return "%I64d";
-#endif
-  } else if (dt->is_primitive(PrimitiveTypeID::u64)) {
-#if defined(TI_PLATFORM_UNIX)
-    return "%llu";
-#else
-    return "%I64u";
-#endif
-  } else if (dt->is_primitive(PrimitiveTypeID::f32)) {
-    return "%f";
-  } else if (dt->is_primitive(PrimitiveTypeID::f64)) {
-    return "%.12f";
-  } else if (dt->is<CustomIntType>()) {
-    return "%d";
-  } else {
-    TI_NOT_IMPLEMENTED
-  }
-}
-
-int data_type_size(DataType t) {
-  // TODO:
-  //  1. Ensure in the old code, pointer attributes of t are correct (by
-  //  setting a loud failure on pointers);
-  //  2. Support pointer types here.
-  t.set_is_pointer(false);
-  if (false) {
-  } else if (t->is_primitive(PrimitiveTypeID::f16))
-    return 2;
-  else if (t->is_primitive(PrimitiveTypeID::gen))
-    return 0;
-  else if (t->is_primitive(PrimitiveTypeID::unknown))
-    return -1;
-
-#define REGISTER_DATA_TYPE(i, j) \
-  else if (t->is_primitive(PrimitiveTypeID::i)) return sizeof(j)
-
-  REGISTER_DATA_TYPE(f32, float32);
-  REGISTER_DATA_TYPE(f64, float64);
-  REGISTER_DATA_TYPE(i8, int8);
-  REGISTER_DATA_TYPE(i16, int16);
-  REGISTER_DATA_TYPE(i32, int32);
-  REGISTER_DATA_TYPE(i64, int64);
-  REGISTER_DATA_TYPE(u8, uint8);
-  REGISTER_DATA_TYPE(u16, uint16);
-  REGISTER_DATA_TYPE(u32, uint32);
-  REGISTER_DATA_TYPE(u64, uint64);
-
-#undef REGISTER_DATA_TYPE
-  else {
-    TI_NOT_IMPLEMENTED
-  }
-}
-
-int data_type_bits(DataType t) {
-  return data_type_size(t) * 8;
-}
-
-std::string data_type_name(DataType t) {
-  if (!t->is<PrimitiveType>()) {
-    return t->to_string();
-  }
-
-  // Handle primitive types below.
-
-  if (false) {
-  }
-#define PER_TYPE(i) else if (t->is_primitive(PrimitiveTypeID::i)) return #i;
-#include "taichi/inc/data_type.inc.h"
-#undef PER_TYPE
-  else
-    TI_NOT_IMPLEMENTED
-}  // namespace lang
-
 std::string snode_type_name(SNodeType t) {
   switch (t) {
 #define PER_SNODE(i) \
@@ -317,71 +235,6 @@ bool command_exist(const std::string &command) {
     return true;
   }
 #endif
-}
-
-namespace {
-class TypePromotionMapping {
- public:
-  TypePromotionMapping() {
-#define TRY_SECOND(x, y)                                   \
-  mapping[std::make_pair(get_primitive_data_type<x>(),     \
-                         get_primitive_data_type<y>())] =  \
-      get_primitive_data_type<decltype(std::declval<x>() + \
-                                       std::declval<y>())>();
-#define TRY_FIRST(x)      \
-  TRY_SECOND(x, float32); \
-  TRY_SECOND(x, float64); \
-  TRY_SECOND(x, int8);    \
-  TRY_SECOND(x, int16);   \
-  TRY_SECOND(x, int32);   \
-  TRY_SECOND(x, int64);   \
-  TRY_SECOND(x, uint8);   \
-  TRY_SECOND(x, uint16);  \
-  TRY_SECOND(x, uint32);  \
-  TRY_SECOND(x, uint64);
-
-    TRY_FIRST(float32);
-    TRY_FIRST(float64);
-    TRY_FIRST(int8);
-    TRY_FIRST(int16);
-    TRY_FIRST(int32);
-    TRY_FIRST(int64);
-    TRY_FIRST(uint8);
-    TRY_FIRST(uint16);
-    TRY_FIRST(uint32);
-    TRY_FIRST(uint64);
-  }
-  DataType query(DataType x, DataType y) {
-    auto primitive =
-        mapping[std::make_pair(to_primitive_type(x), to_primitive_type(y))];
-    return TypeFactory::get_instance().get_primitive_type(primitive);
-  }
-
- private:
-  std::map<std::pair<PrimitiveTypeID, PrimitiveTypeID>, PrimitiveTypeID>
-      mapping;
-  static PrimitiveTypeID to_primitive_type(DataType d) {
-    if (d->is<PointerType>()) {
-      d = d->as<PointerType>()->get_pointee_type();
-      TI_WARN("promoted_type got a pointer input.");
-    }
-
-    if (d->is<VectorType>()) {
-      d = d->as<VectorType>()->get_element_type();
-      TI_WARN("promoted_type got a vector input.");
-    }
-
-    auto primitive = d->cast<PrimitiveType>();
-    TI_ASSERT_INFO(primitive, "Failed to get primitive type from {}",
-                   d->to_string());
-    return primitive->type;
-  };
-};
-TypePromotionMapping type_promotion_mapping;
-}  // namespace
-
-DataType promoted_type(DataType a, DataType b) {
-  return type_promotion_mapping.query(a, b);
 }
 
 std::string TypedConstant::stringify() const {
